pyproject.toml
[tool.poetry]
name = "yolov8-vhr10-finetune"
version = "0.1.0"
description = "YOLOv8 fine-tuning on VHR-10 remote sensing dataset"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "yolov8_finetune", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.0.0"
torchvision = "^0.15.0"
transformers = "^4.35.0"
ultralytics = "^8.0.0"
datasets = "^2.14.0"
pillow = "^10.0.0"
numpy = "^1.24.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
scikit-learn = "^1.3.0"
opencv-python = "^4.8.0"
tqdm = "^4.65.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"